Add confirmation on file delete [Finishes #159144230]

+++ src/app/file/file-list/file-list.component.ts
@@ -206,6 +206,10 @@ export class FileListComponent implements OnInit, OnDestroy {
     }
 
     private removeFile(fileName: string): void {
+
+        if (!confirm('Do you really want to delete ' + fileName + '?')) {
+            return;
+        }
         this.fileService
             .removeFile(this.path.absolutePath(fileName))
             .takeUntil(this.ngUnsubscribe)
